moved comparisons dir into src
